KilterAPI: Update the tables after each fetch from the server
Aggregating the fetched data and then updating was causing intermittent race condition issues where a climb was updated while we were syncing, leading to duplicates in the dataframe index. This caused the update process to fail. Updating the tables as we fetch them is more work but avoids this issue.